add -f so can keep files across successive runs
SCCS-vsn: 5.14

+++ lib/libc/db/test/dbtest.c
@@ -38,7 +38,7 @@ char copyright[] =
 #endif /* not lint */
 
 #ifndef lint
-static char sccsid[] = "@(#)dbtest.c	5.13 (Berkeley) 04/29/93";
+static char sccsid[] = "@(#)dbtest.c	5.14 (Berkeley) 05/01/93";
 #endif /* not lint */
 
 #include <sys/param.h>
@@ -92,11 +92,15 @@ main(argc, argv)
 	DBT data, key, keydata;
 	size_t len;
 	int ch;
-	char *infoarg, *p, buf[8 * 1024];
+	char *fname, *infoarg, *p, buf[8 * 1024];
 
 	infoarg = NULL;
-	while ((ch = getopt(argc, argv, "i:o:")) != EOF)
+	fname = NULL;
+	while ((ch = getopt(argc, argv, "f:i:o:")) != EOF)
 		switch(ch) {
+		case 'f':
+			fname = optarg;
+			break;
 		case 'i':
 			infoarg = optarg;
 			break;
@@ -131,10 +135,13 @@ main(argc, argv)
 			if (*p != '\0')
 				infop = setinfo(type, p);
 
-#define	BACKINGFILE	"/tmp/__dbtest"
 	/* Open the DB. */
-	(void)unlink(BACKINGFILE);
-	if ((dbp = dbopen(BACKINGFILE,
+#define	BACKINGFILE	"/tmp/__dbtest"
+	if (fname == NULL) {
+		fname = BACKINGFILE;
+		(void)unlink(BACKINGFILE);
+	}
+	if ((dbp = dbopen(fname,
 	    O_CREAT | O_RDWR, S_IRUSR | S_IWUSR, type, infop)) == NULL)
 		err("dbopen: %s", strerror(errno));
 	XXdbp = dbp;
@@ -274,6 +281,8 @@ lkey:			switch(command) {
 			    p, lineno);
 		}
 	}
+	if (dbp->close(dbp))
+		err("db->close: %s", strerror(errno));
 	(void)close(ofd);
 	exit(0);
 }
@@ -608,7 +617,7 @@ void
 usage()
 {
 	(void)fprintf(stderr,
-	    "usage: dbtest [-i info] [-o file] type script\n");
+	    "usage: dbtest [-f file] [-i info] [-o file] type script\n");
 	exit(1);
 }
